# docs/contracts/openapi-v1-2026-02-19-phase5-dev.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "movement-whatsapp-automation-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/invoices": {
      "get": {
        "tags": [
          "Admin - Invoices"
        ],
        "summary": "List Invoices",
        "operationId": "list_invoices_api_v1_admin_invoices_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          },
          {
            "name": "therapist_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Therapist Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(pending|issued|voided)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceListItem"
                  },
                  "title": "Response List Invoices Api V1 Admin Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Admin - Invoices"
        ],
        "summary": "Get Invoice Detail",
        "operationId": "get_invoice_detail_api_v1_admin_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invoices/generate": {
      "post": {
        "tags": [
          "Admin - Invoices"
        ],
        "summary": "Generate Invoice",
        "operationId": "generate_invoice_api_v1_admin_invoices_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/therapist/invoices": {
      "get": {
        "tags": [
          "Therapist Invoices"
        ],
        "summary": "List Therapist Invoices",
        "operationId": "list_therapist_invoices_api_v1_therapist_invoices_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(pending|issued|voided)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistInvoiceListItem"
                  },
                  "title": "Response List Therapist Invoices Api V1 Therapist Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/therapist/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Therapist Invoices"
        ],
        "summary": "Get Therapist Invoice Detail",
        "operationId": "get_therapist_invoice_detail_api_v1_therapist_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistInvoiceDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvoiceDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "payment_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Mode"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "special_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Notes"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "issued_by_user_id": {
            "type": "integer",
            "title": "Issued By User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "session_id",
          "amount_cents",
          "currency",
          "description",
          "pdf_url",
          "status",
          "created_at",
          "issued_by_user_id"
        ],
        "title": "InvoiceDetailResponse",
        "description": "Admin invoice detail response."
      },
      "InvoiceGenerateRequest": {
        "properties": {
          "client_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Client Id"
          },
          "session_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Session Id"
          },
          "amount_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 3,
            "title": "Currency",
            "default": "HKD"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description"
          },
          "payment_mode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Mode"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "special_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Notes"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "session_id",
          "description"
        ],
        "title": "InvoiceGenerateRequest",
        "description": "Admin request payload to generate a single-session invoice.",
        "example": {
          "amount_cents": 65000,
          "client_id": 101,
          "currency": "HKD",
          "description": "Physio session invoice",
          "diagnosis": "Bilateral plantar fasciitis",
          "payment_mode": "cash",
          "session_id": 501,
          "special_notes": "Please submit to insurer within 30 days."
        }
      },
      "InvoiceListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "payment_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Mode"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "special_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Notes"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "session_id",
          "amount_cents",
          "currency",
          "description",
          "pdf_url",
          "status",
          "created_at"
        ],
        "title": "InvoiceListItem",
        "description": "Invoice item returned by admin invoice list."
      },
...
      "TherapistInvoiceDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "payment_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Mode"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "special_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Notes"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "client_phone_e164": {
            "type": "string",
            "title": "Client Phone E164"
          },
          "issued_by_user_id": {
            "type": "integer",
            "title": "Issued By User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "session_id",
          "amount_cents",
          "currency",
          "description",
          "pdf_url",
          "status",
          "created_at",
          "client_name",
          "client_phone_e164",
          "issued_by_user_id"
        ],
        "title": "TherapistInvoiceDetailResponse",
        "description": "Therapist-scoped invoice detail response."
      },
      "TherapistInvoiceListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "payment_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Mode"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "special_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Notes"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "client_phone_e164": {
            "type": "string",
            "title": "Client Phone E164"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_id",
          "session_id",
          "amount_cents",
          "currency",
          "description",
          "pdf_url",
          "status",
          "created_at",
          "client_name",
          "client_phone_e164"
        ],
        "title": "TherapistInvoiceListItem",
        "description": "Therapist-scoped invoice list item."
      },
...
    }
  }
}